[ELY-2440] JaasSecurityRealmTest move
