+++ auth/realm/base/src/test/resources/org/wildfly/security/auth/realm/jaas-login.config
@@ -1,0 +1,6 @@
+Entry1 {
+    org.wildfly.security.auth.realm.TestLoginModule required;
+};
+Entry2 {
+    org.wildfly.security.auth.realm.TestLoginModule2 required;
+};
