+++ auth/realm/base/src/test/resources/org/wildfly/security/auth/realm/jaas-login2.config
@@ -1,0 +1,3 @@
+Entry1 {
+    org.wildfly.security.auth.realm.TestLoginModule2 required;
+};
