--- tests/base/src/test/resources/org/wildfly/security/auth/jaas-login.config
@@ -1,6 +1,0 @@
-Entry1 {
-    org.wildfly.security.auth.TestLoginModule required;
-};
-Entry2 {
-    org.wildfly.security.auth.TestLoginModule2 required;
-};
