--- tests/base/src/test/resources/org/wildfly/security/auth/jaas-login2.config
@@ -1,3 +1,0 @@
-Entry1 {
-    org.wildfly.security.auth.TestLoginModule2 required;
-};
